# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "."
  version: "1.0.0"
  title: "Swagger Petstore"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
tags:
- name: "user"
  description: "洗濯物依頼者用"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
- name: "career"
  description: "運搬者用"
schemes:
- "http"
paths:
  /api/user/login:
    get:
      tags:
      - "user"
      summary: "login"
      description: "ログイン"
      operationId: "login"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: ""
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Login"
  /api/user/task:
    put:
      tags:
      - "user"
      summary: "add task"
      description: "洗濯物登録"
      operationId: "addTask"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/task"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Task"
  /api/task/{taskId}:
    get:
      tags:
      - "user"
      summary: "Find task by ID"
      description: "タスク状態検索"
      operationId: "getTaskStatus"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "task id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Pet not found"
      x-swagger-router-controller: "Task"
  /api/career/login:
    get:
      tags:
      - "career"
      summary: "login"
      description: "ログイン"
      operationId: "careerLogin"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: ""
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Login"
  /api/career/task:
    get:
      tags:
      - "career"
      summary: "search task"
      description: "運搬依頼検索"
      operationId: "taskSearch"
      produces:
      - "application/json"
      parameters:
      - name: "careerId"
        in: "query"
        description: ""
        required: true
        type: "string"
      - name: "latitude"
        in: "query"
        description: "緯度"
        required: true
        type: "number"
      - name: "longitude"
        in: "query"
        description: "経度"
        required: true
        type: "number"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid status value"
      x-swagger-router-controller: "Career"
  /api/career/declare:
    post:
      tags:
      - "career"
      summary: "declare task"
      description: "タスク実行を宣言する"
      operationId: "declareTask"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/declare"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Career"
  /api/career/location:
    post:
      tags:
      - "career"
      summary: "regist location"
      description: "位置登録"
      operationId: "registerLocation"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/location"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Career"
  /api/career/receipt/{taskId}:
    post:
      tags:
      - "career"
      summary: "receipt"
      description: "荷物受領"
      operationId: "receipt"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "task id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Career"
  /api/career/start/{taskId}:
    post:
      tags:
      - "career"
      summary: "start"
      description: "洗濯スタート"
      operationId: "start"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "task id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Career"
  /api/career/checkout/{taskId}:
    post:
      tags:
      - "career"
      summary: "checkout"
      description: "洗濯回収"
      operationId: "checkout"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "task id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Career"
  /api/career/end/{taskId}:
    post:
      tags:
      - "career"
      summary: "end task"
      description: "洗濯物引き渡し"
      operationId: "end"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "task id"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      x-swagger-router-controller: "Career"
definitions:
  task:
    type: "object"
    properties:
      userId:
        type: "integer"
        format: "int64"
      plan:
        type: "string"
      objects:
        type: "array"
        items:
          $ref: "#/definitions/taskitem"
  taskitem:
    properties:
      kind:
        type: "string"
      count:
        type: "integer"
  declare:
    type: "object"
    properties:
      careerId:
        type: "integer"
      taskId:
        type: "integer"
  location:
    type: "object"
    properties:
      latitude:
        type: "number"
      longitude:
        type: "number"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"